refactor util.py to util
Adding the FIFO to util is going to make the module cumbersome. Break it into its parts so that it can be simpler to understand what is related and what is not.
